+++ casm/tools/convert/commands/__init__.py
@@ -1,0 +1,1 @@
+"""The casm-convert commands"""